chatbot-3d printer/requirements.txt
adjustText==0.8
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
fastjsonschema==2.16.3
filelock==3.12.4
Fiona==1.9.3
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
GeoAlchemy2==0.13.2
geopandas==0.12.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.22.0
ipython==8.13.1
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langchain==0.0.306
langsmith==0.0.41
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
node==1.2.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.24.3
odict==1.9.0
onnxruntime==1.16.0
openai==0.27.8
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
panda==0.3.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.3.0
plumber==1.7
posthog==3.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.6
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pygame==2.1.2
Pygments==2.15.1
pymongo==4.5.0
pyparsing==3.0.9
PyPika==0.48.9
pyproj==3.5.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.8.8
requests==2.29.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.1
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
TBB==0.2
tenacity==8.2.3
terminado==0.17.1
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.0
tomli==2.0.1
tornado==6.3.1
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.7
yarl==1.9.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==5.0
zope.proxy==5.0.0
